backend-ec2/requirements.txt
tornado==4.2
motor==0.4.1
pyyaml==3.11
jsonschema=2.5.1